Closed group fixes. (#816)
* Fix group updates not syning

* Fix leaving closed groups

* Fix incorrect members being shown on create group dialog

* Linting

* Fix create closed group showing our own conversation

+++ js/conversation_controller.js
@@ -161,26 +161,32 @@
       if (!conversation) {
         return;
       }
-      if (conversation.isPublic()) {
+
+      // Close group leaving
+      if (conversation.isClosedGroup()) {
+        await conversation.leaveGroup();
+      } else if (conversation.isPublic()) {
         const channelAPI = await conversation.getPublicSendData();
         if (channelAPI === null) {
           log.warn(`Could not get API for public conversation ${id}`);
         } else {
           channelAPI.serverAPI.partChannel(channelAPI.channelId);
         }
+      } else if (conversation.isPrivate()) {
+        const deviceIds = await textsecure.storage.protocol.getDeviceIds(id);
+        await Promise.all(
+          deviceIds.map(deviceId => {
+            const address = new libsignal.SignalProtocolAddress(id, deviceId);
+            const sessionCipher = new libsignal.SessionCipher(
+              textsecure.storage.protocol,
+              address
+            );
+            return sessionCipher.deleteAllSessionsForDevice();
+          })
+        );
       }
+
       await conversation.destroyMessages();
-      const deviceIds = await textsecure.storage.protocol.getDeviceIds(id);
-      await Promise.all(
-        deviceIds.map(deviceId => {
-          const address = new libsignal.SignalProtocolAddress(id, deviceId);
-          const sessionCipher = new libsignal.SessionCipher(
-            textsecure.storage.protocol,
-            address
-          );
-          return sessionCipher.deleteAllSessionsForDevice();
-        })
-      );
       await window.Signal.Data.removeConversation(id, {
         Conversation: Whisper.Conversation,
       });
